refac: Move Enum from db to domain and Refactor package structure
