+++ src/main/java/net/causw/adapter/db/CircleRepository.java
@@ -1,8 +1,8 @@
-package net.causw.infra;
+package net.causw.adapter.db;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PostRepository extends JpaRepository<Post, String> {
+public interface CircleRepository extends JpaRepository<Circle, String> {
 }
