+++ src/main/java/net/causw/adapter/db/PostRepository.java
@@ -1,8 +1,8 @@
-package net.causw.infra;
+package net.causw.adapter.db;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BoardRepository extends JpaRepository<Board, String> {
+public interface PostRepository extends JpaRepository<Post, String> {
 }
