+++ src/main/java/net/causw/adapter/db/UserRepository.java
@@ -1,4 +1,4 @@
-package net.causw.infra;
+package net.causw.adapter.db;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
