+++ src/main/java/net/causw/adapter/web/PostController.java
@@ -1,7 +1,7 @@
-package net.causw.web;
+package net.causw.adapter.web;
 
-import net.causw.application.BoardService;
-import net.causw.application.dto.BoardDetailDto;
+import net.causw.application.PostService;
+import net.causw.application.dto.PostDetailDto;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -10,17 +10,17 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/v1/boards")
-public class BoardController {
-    private final BoardService boardService;
+@RequestMapping("/api/v1/posts")
+public class PostController {
+    private final PostService postService;
 
-    public BoardController(BoardService boardService) {
-        this.boardService = boardService;
+    public PostController(PostService postService) {
+        this.postService = postService;
     }
 
     @GetMapping(value = "/{id}")
     @ResponseStatus(value = HttpStatus.OK)
-    public BoardDetailDto findById(@PathVariable String id) {
-        return this.boardService.findById(id);
+    public PostDetailDto findById(@PathVariable String id) {
+        return this.postService.findById(id);
     }
 }
