+++ src/main/java/net/causw/application/dto/UserAuthDto.java
@@ -2,7 +2,7 @@ package net.causw.application.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import net.causw.infra.UserAuth;
+import net.causw.adapter.db.UserAuth;
 
 import java.time.LocalDateTime;
