+++ src/main/java/net/causw/domain/model/Role.java
@@ -1,4 +1,4 @@
-package net.causw.infra;
+package net.causw.domain.model;
 
 import lombok.Getter;
